feat: break the license into another file and include it in the indexer
