+++ package.json
@@ -71,5 +71,6 @@
         "vitest": "2.1.3"
       }
     }
-  }
+  },
+  "type": "module"
 }
